LayoutTests/fast/canvas/CanvasRenderingContext2DSettings-pixelFormat-disabled-expected.txt

Test that the colorType member of CanvasRenderingContext2DSettings is unsupported when CanvasColorTypeEnabled=false.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testSettings.colorType is undefined.
PASS testSettings.colorType is undefined.
PASS testSettings.colorType is undefined.
PASS testSettings.colorType is undefined.
PASS successfullyParsed is true

TEST COMPLETE
